requirements.txt
Django==5.2.4
django_environ==0.12.0
djangorestframework==3.16.0
djangorestframework-simplejwt==5.5.1